# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Property Management API",
    "version": "1.0.0",
    "description": "API documentation for property management system"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["tenant", "landlord", "admin"] },
          "phoneNumber": { "type": "string" },
          "idNumber": { "type": "string" }
        }
      },
      "Unit": {
        "type": "object",
        "properties": {
          "unitName": { "type": "string" },
          "rentAmount": { "type": "number" },
          "isOccupied": { "type": "boolean" },
          "paymentStatus": { "type": "string", "enum": ["paid", "unpaid"] }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "role", "phoneNumber", "idNumber"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["tenant", "landlord", "admin"] },
                  "phoneNumber": { "type": "string" },
                  "idNumber": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/units": {
      "get": {
        "tags": ["Units"],
        "summary": "Get all units",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of units",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Unit" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Units"],
        "summary": "Create a new unit",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Unit" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Unit created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Unit" }
              }
            }
          }
        }
      }
    }
  }
}